requirements.txt
deta==1.1.0
pandas==1.4.3
python-dotenv==0.20.0
PyYAML==6.0
streamlit==1.10.0
streamlit-authenticator==0.1.5